Added fake tests for attributes extraction
