+++ xds/src/main/java/io/grpc/xds/internal/rbac/engine/AuthorizationEngine.java
@@ -119,10 +119,8 @@ public class AuthorizationEngine<ReqT, RespT> {
    * Determines whether a gRPC call is allowed, denied, or unable to decide.
    * @param args evaluate argument that is used to evaluate the RBAC conditions.
    * @return an AuthorizationDecision generated by CEL Evaluation Engine.
-   * @throws InterpreterException if something goes wrong in CEL library.
    */
-  public AuthorizationDecision evaluate(EvaluateArgs<ReqT, RespT> args) 
-      throws IllegalArgumentException {
+  public AuthorizationDecision evaluate(EvaluateArgs<ReqT, RespT> args) {
     List<String> unknownPolicyNames = new ArrayList<>();
     // Set up activation used in CEL library's eval function.
     Activation activation = Activation.copyOf(extractFields(args));
